props passed with handlers and state
